Moved some classes to another package
